+++ src/config_repairs.h
@@ -10,7 +10,8 @@ namespace ild
 enum class ConfigRepair { mechanic_dialog, prince_dialog, skill_text, info_root, camp_condition,
     burer_dialog, psi_sound, car_text, skat_upgrade, mutant_sounds, vodka_task, counter_wide,
     counter_normal, trader_refusal, bronevik_profile, detector_task, detector_text, guide_shotgun,
-    monolith_task, abram_pistol, samogon_task, killer_surrender, propysk_text, skill_banner };
+    monolith_task, abram_pistol, samogon_task, killer_surrender, propysk_text, skill_banner, kuzma_timeout,
+    tunnel_input };
 struct ConfigRepairSource
 {
     std::wstring_view relative;